fuir: Make FUIR Abstract, move implementation to new class AirFUIR
First step towards adding a FUIR that is based on fum files directly.
